added rotation of faces on initialization
This code is broken. It is possible that I have confused myself, and
that I am passing references of static objects, that get transformed
back in their parent class, even when manipulating said references in a
separate class. If this would turn out to be the case, I could make this
whole program much simpler. Maybe the UI will help in debugging this.